Rename `tuntap` package to `tun`
We haven't had TAP support in ages.
