+++ src/tun/tun_bsd.go
@@ -1,7 +1,7 @@
 //go:build openbsd || freebsd
 // +build openbsd freebsd
 
-package tuntap
+package tun
 
 import (
 	"encoding/binary"
